move non program specific stuff to common folder
organise the code base some more by putting common stuff in the common folder
