+++ source/common/common.h
@@ -1,0 +1,10 @@
+#pragma once
+
+#include <inttypes.h>
+#include <stddef.h>
+#include <stdbool.h>
+#include <string.h>
+#include <ctype.h>
+#include <stdlib.h>
+#include <stdio.h>
+#include <types.h>
