+++ source/common/types.h
